+++ unity-client/Assets/Scripts/MainScripts/DCL/Controllers/CharacterController/Attachable.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c65f8fb1a8d804ad39b0d1d176e1b864
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
