+++ unity-client/Assets/Scripts/MainScripts/DCL/Controllers/CharacterController/Attachable/Attachable.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 9ba582ddc25eb45d6a166c32bda1d1c0
+guid: 658e13c0ca28b4de8aafc4947b1fa2f8
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
