+++ unity-client/Assets/Scripts/MainScripts/DCL/Controllers/CharacterController/Attachable/AvatarReference.cs
@@ -3,11 +3,11 @@ using DCL.Models;
 using DCL.Controllers;
 using System.Collections.Generic;
 
-public class PlayerReference : MonoBehaviour
+public class AvatarReference : MonoBehaviour
 {         
 
-    public static PlayerReference i { get; private set; }    
-    // From Gameobject instance id to its entity
+    public static AvatarReference i { get; private set; }    
+    // From Gameobject instance id to its entity representation
     private readonly Dictionary<int, DecentralandEntity> entities = new Dictionary<int, DecentralandEntity>();
 
     public void AttachEntity(DecentralandEntity entity)
@@ -15,8 +15,7 @@ public class PlayerReference : MonoBehaviour
         int instanceId = entity.gameObject.GetInstanceID();
         entities.Add(instanceId, entity);
         entity.gameObject.transform.SetParent(transform, false);
-        entity.OnRemoved += OnEntityRemoval;
-        Debug.Log("Attaching to player");
+        entity.OnRemoved += OnEntityRemoval;        
     }
 
     private void Awake()
@@ -29,15 +28,9 @@ public class PlayerReference : MonoBehaviour
         }
 
         i = this;
-
-        // Listen to changes on the camera mode
-        CommonScriptableObjects.cameraMode.OnChange += OnCameraModeChange;
-
+        
         // Listen to changes on the player position
-        CommonScriptableObjects.playerWorldPosition.OnChange += OnPlayerPositionChange;
-
-        // Trigger the initial camera set
-        OnCameraModeChange(CommonScriptableObjects.cameraMode, CommonScriptableObjects.cameraMode);
+        CommonScriptableObjects.playerWorldPosition.OnChange += OnPlayerPositionChange;        
     }        
 
     private void OnPlayerPositionChange(Vector3 newPosition, Vector3 prev)
@@ -48,25 +41,7 @@ public class PlayerReference : MonoBehaviour
             DecentralandEntity entity = entities[childId];
             DCL.SceneController.i.boundariesChecker.AddEntityToBeChecked(entity);        
         }        
-    }
-
-    private void UpdateForward(Vector3 newFordward, Vector3 prev)
-    {       
-        transform.forward = newFordward;
-    }   
-
-    private void OnCameraModeChange(CameraMode.ModeId newMode, CameraMode.ModeId prev)
-    {
-        if (newMode == CameraMode.ModeId.ThirdPerson)
-        {
-            CommonScriptableObjects.cameraForward.OnChange -= UpdateForward;
-            transform.forward = transform.parent.forward;
-        }
-        else
-        {
-            CommonScriptableObjects.cameraForward.OnChange += UpdateForward;
-        }
-    }
+    }         
 
     private void OnEntityRemoval(DecentralandEntity entity)
     {
@@ -74,9 +49,7 @@ public class PlayerReference : MonoBehaviour
     }
 
     private void OnDestroy()
-    {
-        CommonScriptableObjects.cameraMode.OnChange -= OnCameraModeChange;
-        CommonScriptableObjects.cameraForward.OnChange -= UpdateForward;
+    {        
         CommonScriptableObjects.playerWorldPosition.OnChange -= OnPlayerPositionChange;
     }
 }
