+++ unity-client/Assets/Scripts/MainScripts/DCL/Controllers/CharacterController/Attachable/AvatarReference.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 52f3d753c04084bb1bf9081db4e7398f
+guid: a1122e58ceb994b0da9b817a2a002120
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
