+++ unity-client/Assets/Scripts/MainScripts/DCL/Controllers/CharacterController/Attachable/PlayerReference.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8763a10505c8f4ca5847a3292898e1fd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
